Override the iOS hardware mute switch.
Yeesh... It turns out that (as of 2019-08-20), on iOS devices, HTML5 audio is "locked" until a sound is played through it during a user input event callback, and, in that locked state, WebAudio is muted when the hardware mute switch is on. Override the switch by playing some HTML5 audio the first time the top window is touched.